speech/requirements.txt
astroid==2.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
cursor==1.2.0
dialogflow==0.5.2
enum34==1.1.6
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-speech==0.36.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
log-symbols==0.0.11
mccabe==0.6.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
PyAudio==0.2.11
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.11.0
termcolor==1.1.0
typed-ast==1.2.0
urllib3==1.24.1
webrtcvad==2.0.10
wrapt==1.11.1
